Cleans up types
